module7-task1: - рефакторинг компонента comment and cities
